Crate ktracepoint

A Rust library for defining and managing tracepoints in a no_std environment. It provides macros and structures to create tracepoints, manage their state, and handle trace events efficiently. The library is designed to be lightweight and suitable for embedded systems or kernel-level programming where the standard library is not available. It leverages Rust’s powerful macro system to simplify the creation and management of tracepoints. The macros provided by this library allow for easy insertion of tracepoints into code with minimal overhead.

Re-exports

pub use tp_lexer;

Macros

define_event_trace

Define a tracepoint with the given parameters.

paste

Structs

CommonTracePointMeta

CommonTracePointMeta holds metadata for a common tracepoint.

EventInfo

EventInfo holds information about a specific trace event.

EventsSubsystem

EventsSubsystem represents a collection of events under a specific subsystem.

KernelCodeManipulator

A utility struct to manipulate kernel code, primarily used for ensuring that we can modify kernel code safely.

TraceCmdLineCache

A cache for storing command line arguments for each trace point.

TraceCmdLineCacheSnapshot

A snapshot of the command line cache at a specific point in time.

TraceEntry

A trace entry structure that holds metadata about a trace event.

TraceEntryParser

A parser for trace entries that formats them into human-readable strings.

TraceFilterFile

TraceFilterFile provides a way to set filters on the tracepoint.

TracePipeRaw

A raw trace pipe buffer that stores trace events as byte vectors.

TracePipeSnapshot

A snapshot of the trace pipe buffer at a specific point in time.

TracePoint

The TracePoint structure represents a tracepoint in the system.

TracePointEnableFile

TracePointEnableFile provides a way to enable or disable the tracepoint.

TracePointFormatFile

TracePointFormatFile provides a way to get the format of the tracepoint.

TracePointFunc

A structure representing a registered tracepoint callback function.

TracePointIdFile

TracePointEnableFile provides a way to enable or disable the tracepoint.

TracePointMap

TracePointMap is a mapping from tracepoint IDs to TracePoint references.

TracingEventsManager

TracingEventsManager manages tracing events, subsystems, and tracepoints.

Traits

AsU64

A trait to convert various types to u64 representation.

KernelTraceOps

KernelTraceOps trait provides kernel-level operations for tracing.

RawTracePointCallBackFunc

A trait for raw callback functions that can be registered with a tracepoint.

TracePipeOps

A trait defining operations for a trace pipe buffer.

TracePointCallBackFunc

A trait for callback functions that can be registered with a tracepoint.

Functions

global_init_events

Initialize the tracing events