ktracepoint

Macro define_event_trace

macro_rules! define_event_trace {
    (
        $name:ident,
        TP_lock($lock:path),
        TP_kops($kops:path),
        TP_system($system:ident),
        TP_PROTO($($arg:ident:$arg_type:ty),+ $(,)?),
        TP_STRUCT__entry{$($entry:ident:$entry_type:ty),+ $(,)?},
        TP_fast_assign{$($assign:ident:$value:expr),+ $(,)?},
        TP_ident($tp_ident:ident),
        TP_printk($fmt_expr: expr)
    ) => { ... };
}

Define a tracepoint with the given parameters.

This macro generates a tracepoint with the specified name, arguments, entry structure, assignment logic, identifier, and print format.

Parameters

name: The name of the tracepoint.

TP_lock: The lock type to use for the tracepoint.

TP_kops: The kernel trace operations type. [crate::KernelTraceOps] is expected to be implemented for this type.

TP_system: The subsystem or system to which the tracepoint belongs.

TP_PROTO: The prototype of the tracepoint function.

TP_STRUCT__entry: The structure of the tracepoint entry. WARN: User need to make sure the layout of the struct is compatible with C layout.

TP_fast_assign: The assignment logic for the tracepoint entry.

TP_ident: The identifier for the tracepoint entry.

TP_printk: The print format for the tracepoint.

Example

use lock_api::Mutex;
use crate::KernelTraceOps;
define_event_trace!(
    Mutex<()>,
    Kops,
    TEST2,
    TP_PROTO(a: u32, b: u32),
    TP_STRUCT__entry{
          a: u32,
          b: u32,
    },
    TP_fast_assign{
          a:a,
          b:{
            // do something with b
            b
          }
    },
    TP_ident(__entry),
    TP_printk({
          // do something with __entry
          format!("Hello from tracepoint! a={}, b={}", __entry.a, __entry.b)
    })
);