ktracepoint

Struct TracePoint

pub struct TracePoint<L: RawMutex + 'static, K: KernelTraceOps + 'static> { /* private fields */ }

The TracePoint structure represents a tracepoint in the system.

Implementations

impl<L: RawMutex + 'static, K: KernelTraceOps + 'static> TracePoint<L, K>

pub const fn new( key: &'static RawStaticFalseKey<KernelCodeManipulator<K>>, name: &'static str, system: &'static str, fmt_func: fn(&[u8]) -> String, trace_print_func: fn() -> String, schema: Schema, ) -> Self

Creates a new TracePoint instance.

pub fn schema(&self) -> &Schema

Returns the schema of the tracepoint.

pub fn name(&self) -> &'static str

Returns the name of the tracepoint.

pub fn system(&self) -> &'static str

Returns the system of the tracepoint.

pub fn id(&self) -> u32

Returns the ID of the tracepoint.

pub fn flags(&self) -> u8

Returns the flags of the tracepoint.

pub fn set_compiled_expr(&self, compiled: Option<Compiled>)

Sets the compiled expression for the tracepoint.

pub fn get_compiled_expr(&self) -> Option<Compiled>

Returns the compiled expression for the tracepoint.

pub fn print_fmt(&self) -> String

Returns a string representation of the format function for the tracepoint.

You can use cat /sys/kernel/debug/tracing/events/syscalls/sys_enter_openat/format in linux to see the format of the tracepoint.

pub fn register(&self, func: fn(), data: Box<dyn Any + Sync + Send>)

Register a callback function to the tracepoint

pub fn unregister(&self, func: fn())

Unregister a callback function from the tracepoint

pub fn callback_list(&self, f: &dyn Fn(&TracePointFunc))

Iterate over all registered callback functions

pub fn register_event_callback( &self, callback_id: usize, callback: Box<dyn TracePointCallBackFunc>, )

Register a event callback function to the tracepoint

This function will be called when default tracepoint fmt function is called.

pub fn unregister_event_callback(&self, callback_id: usize)

Unregister a event callback function from the tracepoint

pub fn event_callback_list(&self, f: &dyn Fn(&Box<dyn TracePointCallBackFunc>))

Iterate over all registered event callback functions

pub fn register_raw_event_callback( &self, callback_id: usize, callback: Box<dyn RawTracePointCallBackFunc>, )

Register a raw event callback function to the tracepoint

pub fn unregister_raw_event_callback(&self, callback_id: usize)

Unregister a raw event callback function from the tracepoint

pub fn raw_event_callback_list( &self, f: &dyn Fn(&Box<dyn RawTracePointCallBackFunc>), )

Iterate over all registered raw event callback functions

pub fn enable_default(&self)

Enable the tracepoint for the default print

pub fn disable_default(&self)

Disable the tracepoint for the default print

pub fn default_is_enabled(&self) -> bool

Check if the tracepoint is enabled for the default print

pub fn enable_event(&self)

Enable the tracepoint event for custom event handling

pub fn disable_event(&self)

Disable the tracepoint event for custom event handling

pub fn event_is_enabled(&self) -> bool

Check if the tracepoint event is enabled for custom event handling