ktracepoint

Trait TracePipeOps

pub trait TracePipeOps {
    // Required methods
    fn peek(&self) -> Option<&Vec<u8>>;
    fn pop(&mut self) -> Option<Vec<u8>>;
    fn is_empty(&self) -> bool;
}

A trait defining operations for a trace pipe buffer.

Required Methods

fn peek(&self) -> Option<&Vec<u8>>

Returns the first event in the trace pipe buffer without removing it.

fn pop(&mut self) -> Option<Vec<u8>>

Remove and return the first event in the trace pipe buffer.

fn is_empty(&self) -> bool

Whether the trace pipe buffer is empty.